backend/requirements.txt
# -------- Core API --------
fastapi==0.121.2
uvicorn==0.38.0
starlette==0.49.3
slowapi==0.1.9
python-multipart==0.0.20
python-dotenv==1.2.1
orjson==3.11.4
httpx==0.28.1
tenacity==9.1.2
rq==1.16.2
redis==5.0.4

# -------- Database / Vector --------
SQLAlchemy[asyncio]==2.0.36
psycopg[binary]==3.3.2
pgvector==0.4.2
asyncpg
postgrest==2.24.0

# -------- Supabase --------
supabase==2.24.0
supabase-auth==2.24.0
supabase-functions==2.24.0
storage3==2.24.0
realtime==2.24.0

# -------- Google GenAI --------
google-genai==1.6.0

# -------- OpenAI --------
openai>=1.0.0

# -------- LangChain / RAG --------
langchain==1.0.5
langchain-core==1.0.5
langchain-text-splitters==1.0.0
langgraph==1.0.3
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.4
langgraph-sdk==0.2.9
langsmith==0.4.42
pypdf==6.2.0

# -------- Security / Auth --------
cryptography==46.0.3
PyJWT==2.10.1
rsa==4.9.1

# -------- Utils --------
nltk==3.9.1
psutil==7.1.3
requests==2.32.5
requests-toolbelt==1.0.0
tqdm==4.67.1
xxhash==3.6.0
zstandard==0.25.0
PyYAML==6.0.3
packaging==25.0

# -------- Testing --------
pytest
pytest-asyncio
anyio==4.11.0

# -------- Typing / Validation --------
pydantic==2.12.4
pydantic_core==2.41.5
typing_extensions==4.15.0
typing-inspection==0.4.2
annotated-types==0.7.0
